refactor: move create app handler back to v1
