+++ server/src/models/v1/in/CreateAppModel.js
@@ -1,50 +1,30 @@
 const Joi = require('@hapi/joi')
-const { isSemver } = require('../../helpers')
-
 const { AppTypes } = require('../../../enums')
+const { isSemver } = require('../../helpers')
 
 const CreateModelAppData = Joi.object().keys({
     name: Joi.string(),
-    description: Joi.string().allow(''),
+    description: Joi.string(),
     appType: Joi.string().valid(...AppTypes),
-    sourceUrl: Joi.string()
-        .uri()
-        .allow(''),
+    sourceUrl: Joi.string().uri(),
     developer: Joi.object().keys({
         name: Joi.string(),
         email: Joi.string().email(),
-        address: Joi.string().allow(''),
-        organisation: Joi.string(),
+        organisationId: Joi.string(),
     }),
-    versions: Joi.array().items(
-        Joi.object().keys({
-            version: Joi.string().custom(isSemver, 'semver validate'),
-            minDhisVersion: Joi.string(),
-            maxDhisVersion: Joi.string().allow(''),
-            demoUrl: Joi.string()
-                .uri()
-                .allow(''),
-            channel: Joi.string(),
-        })
-    ),
-    images: Joi.array().items(
-        Joi.object({
-            caption: Joi.string().allow('', null),
-            description: Joi.string().allow('', null),
-        })
-    ),
-    owner: Joi.object({
-        email: Joi.string()
-            .email()
-            .required(),
-        name: Joi.string().required(),
+    version: Joi.object().keys({
+        version: Joi.string().custom(isSemver, 'semver validate'),
+        minDhisVersion: Joi.string(),
+        maxDhisVersion: Joi.string().allow(''),
+        demoUrl: Joi.string().uri().allow(''),
+        channel: Joi.string(),
     }),
 })
 
 const payloadSchema = Joi.object({
     //multipart gets parsed as streams so we have to allow any and manually validate in the handler.
     app: Joi.any(),
-    imageFile: Joi.any(),
+    logo: Joi.any(),
     file: Joi.any(),
 })
