+++ server/src/models/v2/in/CreateAppModel.js
@@ -1,34 +1,15 @@
 const Joi = require('@hapi/joi')
-const { isSemver } = require('../../helpers')
-
 const { AppTypes } = require('../../../enums')
 
 const CreateModelAppData = Joi.object().keys({
-    name: Joi.string(),
-    description: Joi.string(),
     appType: Joi.string().valid(...AppTypes),
-    sourceUrl: Joi.string().uri(),
     developer: Joi.object().keys({
-        name: Joi.string(),
-        email: Joi.string().email(),
         organisationId: Joi.string(),
     }),
-    version: Joi.object().keys({
-        version: Joi.string().custom(isSemver, 'semver validate'),
-        minDhisVersion: Joi.string(),
-        maxDhisVersion: Joi.string().allow(''),
-        demoUrl: Joi.string()
-            .uri()
-            .allow(''),
-        channel: Joi.string(),
-    })
 })
 
 const payloadSchema = Joi.object({
-    //multipart gets parsed as streams so we have to allow any and manually validate in the handler.
     app: Joi.any(),
-    logo: Joi.any(),
-    file: Joi.any(),
 })
 
 module.exports = {
